+++ src/gfx/interface/impl/imgui.cpp
@@ -96,8 +96,8 @@ void imgui::new_frame()
     glfwGetFramebufferSize(_gfx_context->window(), &display_w, &display_h);
     io.DisplaySize             = ImVec2(static_cast<float>(w), static_cast<float>(h));
     io.DisplayFramebufferScale = ImVec2(w > 0 ? (static_cast<float>(display_w) / w) : 0, h > 0 ? (static_cast<float>(display_h) / h) : 0);
-    io.DeltaTime               = glfwGetTime() - _last_time;
-    _last_time                 = glfwGetTime();
+    io.DeltaTime               = static_cast<float>(glfwGetTime() - _last_time);
+    _last_time                 = static_cast<float>(glfwGetTime());
 
     if(glfwGetWindowAttrib(_gfx_context->window(), GLFW_FOCUSED))
     {
